# extensions/gemini/openapi.yaml
openapi: 3.0.3
info:
  title: Hegelion Prompt API
  description: >
    Returns Hegelian Thesis → Antithesis → Synthesis prompts/workflows that any LLM can execute.
    No API keys required—the caller runs the prompts with its own model.
  version: "0.2.0"
servers:
  - url: https://hegelion-prompt.example.com
    description: Replace with your Render/Cloud Run URL
paths:
  /dialectical_workflow:
    post:
      summary: Generate a structured dialectical workflow
      operationId: dialecticalWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowRequest"
      responses:
        "200":
          description: Workflow JSON describing thesis → antithesis → synthesis steps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /dialectical_single_shot:
    post:
      summary: Generate one comprehensive dialectical prompt
      operationId: dialecticalSingleShot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SingleShotRequest"
      responses:
        "200":
          description: Prompt string that guides the LLM through all phases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /thesis_prompt:
    post:
      summary: Generate the thesis phase prompt
      operationId: thesisPrompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThesisRequest"
      responses:
        "200":
          description: Thesis prompt specification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /antithesis_prompt:
    post:
      summary: Generate the antithesis (critique) prompt(s)
      operationId: antithesisPrompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AntithesisRequest"
      responses:
        "200":
          description: Either a single prompt or council prompts
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/PromptResponse"
                  - $ref: "#/components/schemas/CouncilPromptResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /synthesis_prompt:
    post:
      summary: Generate the synthesis prompt
      operationId: synthesisPrompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SynthesisRequest"
      responses:
        "200":
          description: Synthesis prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
components:
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    WorkflowRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
        use_search:
          type: boolean
          default: false
        use_council:
          type: boolean
          default: false
        use_judge:
          type: boolean
          default: false
    WorkflowResponse:
      type: object
      properties:
        workflow:
          type: object
          description: JSON workflow with thesis/antithesis/synthesis steps
    SingleShotRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
        use_search:
          type: boolean
          default: false
        use_council:
          type: boolean
          default: false
    ThesisRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
    AntithesisRequest:
      type: object
      required: [query, thesis]
      properties:
        query:
          type: string
        thesis:
          type: string
        use_search:
          type: boolean
          default: false
        use_council:
          type: boolean
          default: false
    SynthesisRequest:
      type: object
      required: [query, thesis, antithesis]
      properties:
        query:
          type: string
        thesis:
          type: string
        antithesis:
          type: string
    PromptResponse:
      type: object
      properties:
        phase:
          type: string
        prompt:
          type: string
        instructions:
          type: string
        expected_format:
          type: string
    CouncilPromptResponse:
      type: object
      properties:
        prompts:
          type: array
          items:
            $ref: "#/components/schemas/PromptResponse"
    ErrorResponse:
      type: object
      properties:
        error:
          type: string